# api/swagger.yaml
definitions:
...
  controller.OrgAcceptMemberRequest:
    properties:
      msg:
        type: string
      org_name:
        type: string
    required:
    - org_name
    type: object
  controller.OrgApproveMemberRequest:
    properties:
      msg:
        type: string
      org_name:
        type: string
      user:
        type: string
    required:
    - org_name
    type: object
  controller.OrgInviteMemberRequest:
    properties:
      msg:
        type: string
      org_name:
        type: string
      role:
        type: string
      user:
        type: string
    required:
    - org_name
    - role
    - user
    type: object
  controller.OrgMemberEditRequest:
    properties:
      role:
        type: string
      user:
        type: string
    required:
    - role
    - user
    type: object
  controller.OrgReqMemberRequest:
    properties:
      msg:
        type: string
      org_name:
        type: string
    required:
    - org_name
    type: object
  controller.OrgRevokeInviteRequest:
    properties:
      msg:
        type: string
      org_name:
        type: string
      user:
        type: string
    required:
    - org_name
    type: object
  controller.OrgRevokeMemberReqRequest:
    properties:
      msg:
        type: string
      org_name:
        type: string
      user:
        type: string
    required:
    - org_name
    type: object
  controller.ResponseData:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
...
  controller.orgBasicInfoUpdateRequest:
    properties:
      allow_request:
        type: boolean
      avatar_id:
        type: string
      default_role:
        type: string
      description:
        type: string
      fullname:
        type: string
      website:
        type: string
    type: object
  controller.orgCreateRequest:
    properties:
      avatar_id:
        type: string
      description:
        type: string
      fullname:
        type: string
      name:
        type: string
      website:
        type: string
    required:
    - name
    type: object
  controller.orgMemberRemoveRequest:
    properties:
      user:
        type: string
    required:
    - user
    type: object
...
info:
  contact: {}
paths:
  /v1/account/{name}:
    get:
      consumes:
      - application/json
      description: get organization or user info
      parameters:
      - description: name of the user of organization
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "404":
          description: user not found
      summary: User or organization info
      tags:
      - Organization
  /v1/invite:
    delete:
      consumes:
      - application/json
      description: Revoke invitation of the organization
      parameters:
      - description: body of the invitation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.OrgRevokeInviteRequest'
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: RevokeInvitation
      tags:
      - Organization
    get:
      consumes:
      - application/json
      description: List invitation of the organization
      parameters:
      - description: organization name
        in: query
        name: org_name
        type: string
      - description: invitee name
        in: query
        name: invitee
        type: string
      - description: inviter name
        in: query
        name: inviter
        type: string
      - description: 'invitation status, can be: pending/approved/rejected'
        in: query
        name: status
        type: string
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: page index
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: ListInvitation
      tags:
      - Organization
    post:
      consumes:
      - application/json
      description: Send invitation to a user to join the organization
      parameters:
      - description: body of the invitation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.OrgInviteMemberRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: InviteMember
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: Accept invite of the organization
      parameters:
      - description: body of the invitation
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.OrgAcceptMemberRequest'
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: AcceptInvite
      tags:
      - Organization
...
  /v1/organization:
    get:
      consumes:
      - application/json
      description: get organization info
      parameters:
      - description: filter by owner
        in: query
        name: owner
        type: string
      - description: filter by username
        in: query
        name: username
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: List
      tags:
      - Organization
    post:
      consumes:
      - application/json
      description: create a new organization
      parameters:
      - description: body of new organization
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.orgCreateRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: Create
      tags:
      - Organization
  /v1/organization/{name}:
    delete:
      consumes:
      - application/json
      description: delete a organization
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Delete
      tags:
      - Organization
    get:
      consumes:
      - application/json
      description: get organization info
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      summary: Get
      tags:
      - Organization
    post:
      consumes:
      - application/json
      description: leave the organization
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: Leave
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: update org basic info
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      - description: body of new organization
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.orgBasicInfoUpdateRequest'
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: Update
      tags:
      - Organization
  /v1/organization/{name}/member:
    delete:
      consumes:
      - application/json
      description: Remove a member from a organization
      parameters:
      - description: body of the removed member
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.orgMemberRemoveRequest'
      - description: name
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - Bearer: []
      summary: RemoveMember
      tags:
      - Organization
    get:
      consumes:
      - application/json
      description: list organization members
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: ListMember
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: Edit a member to the organization's role
      parameters:
      - description: body of new member
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.OrgMemberEditRequest'
      - description: name
        in: path
        name: name
        required: true
        type: string
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: EditMember
      tags:
      - Organization
  /v1/request:
    delete:
      consumes:
      - application/json
      description: Revoke member request of the organization
      parameters:
      - description: body of the member request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.OrgRevokeMemberReqRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: RevokeMember
      tags:
      - Organization
    get:
      consumes:
      - application/json
      description: List requests of the organization
      parameters:
      - description: organization name
        in: query
        name: org_name
        type: string
      - description: invitee name
        in: query
        name: requester
        type: string
      - description: 'invitation status, can be: pending/approved/rejected'
        in: query
        name: status
        type: string
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: page index
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: ListRequests
      tags:
      - Organization
    post:
      consumes:
      - application/json
      description: Request to be a member of the organization
      parameters:
      - description: body of the member request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.OrgReqMemberRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: RequestMember
      tags:
      - Organization
    put:
      consumes:
      - application/json
      description: Approve a user's member request of the organization
      parameters:
      - description: body of the accept
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.OrgApproveMemberRequest'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - Bearer: []
      summary: ApproveRequest
      tags:
      - Organization
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and api Bearer.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"